+++ cashRegister/product.py
@@ -1,4 +1,0 @@
-
-class product:
-    def __init_(code):
-        pass