+++ test/cashRegister/checkout.py
@@ -1,3 +1,0 @@
-class Checkout:
-    def __init__():
-        pass