+++ test/cashRegister/pricing_rules.py
@@ -1,3 +1,0 @@
-class Pricing_rules:
-    def __init__():
-        pass